Update blob cache to decoupled sidecar structure
