+++ beacon_node/beacon_chain/src/lib.rs
@@ -6,7 +6,6 @@ mod beacon_fork_choice_store;
 pub mod beacon_proposer_cache;
 mod beacon_snapshot;
 pub mod blob_cache;
-pub mod blob_sidecar_cache;
 pub mod blob_verification;
 pub mod block_reward;
 mod block_times_cache;
