+++ beacon_node/http_api/src/publish_blobs.rs
@@ -1,4 +1,3 @@
-use beacon_chain::blob_sidecar_cache::BlobSidecarsCache;
 use beacon_chain::{BeaconChain, BeaconChainTypes};
 use lighthouse_network::PubsubMessage;
 use network::NetworkMessage;
@@ -11,20 +10,17 @@ use warp::Rejection;
 /// Handles a request from the HTTP API for blobs.
 pub async fn publish_blob<T: BeaconChainTypes>(
     blinded_blob_sidecar: SignedBlindedBlobSidecar,
-    _chain: Arc<BeaconChain<T>>,
+    chain: Arc<BeaconChain<T>>,
     network_tx: &UnboundedSender<NetworkMessage<T::EthSpec>>,
     log: Logger,
 ) -> Result<(), Rejection> {
-    // FIXME(jimmy) replace fake impl with the real one
-    let blob_cache = BlobSidecarsCache::default();
-
     // TODO check for fork
     let (block_root, blob_index) = (
         &blinded_blob_sidecar.message.blob_root,
         blinded_blob_sidecar.message.index,
     );
 
-    if let Some(full_sidecar) = blob_cache.pop(block_root, blob_index) {
+    if let Some(full_sidecar) = chain.blob_cache.pop(block_root, blob_index) {
         let signed_blob_sidecar = SignedBlobSidecar {
             message: full_sidecar,
             signature: blinded_blob_sidecar.signature,
